Student report generate done
