+++ accounts/api/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 
 from accounts.api.views import (
     StudentQRView,
-    StudentReportGeneratorAPIView,
     UserCreateAPIView,
     UserCreateOTPVerifyAPIView,
     UserDetailAPIView,
@@ -31,9 +30,4 @@ urlpatterns = [
     ),
     path("update/me/", UserUpdateAPIView.as_view(), name="update"),
     path("qr/", StudentQRView.as_view(), name="qrretrive"),
-    path(
-        "report/generate/",
-        StudentReportGeneratorAPIView.as_view(),
-        name="generator-student",
-    ),
 ]
