+++ accounts/api_admin/urls.py
@@ -2,6 +2,7 @@ from django.urls import path
 
 from accounts.api_admin.views import (
     GetSMSCreditAdminAPIView,
+    StudentReportGeneratorAPIView,
     UserCreateAdminAPIView,
     UserListAdminAPIView,
     UserRetrieveAdminAPIView,
@@ -18,4 +19,9 @@ urlpatterns = [
     path("update/<int:pk>/", UserUpdateAdminAPIView.as_view(), name="update"),
     path("sms/credit/", GetSMSCreditAdminAPIView.as_view(), name="credit"),
     path("roles/list/", UserRolesView.as_view(), name="roles-list"),
+    path(
+        "report/generate/",
+        StudentReportGeneratorAPIView.as_view(),
+        name="generator-student",
+    ),
 ]
